fix: change quantity field type to integer in WarehousePosition model and related components

+++ database/migrations/2026_04_27_214249_add_quantity_to_warehouse_positions_table.php
@@ -9,7 +9,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('warehouse_positions', function (Blueprint $table) {
-            $table->decimal('quantity', 10, 3)->nullable()->after('started');
+            $table->integer('quantity')->nullable()->after('started');
         });
     }
